+++ lib/versions/v1/Api/DealsApi.php
@@ -3838,14 +3838,15 @@ class DealsApi
      * @param  int $id The ID of the deal (required)
      * @param  int|0 $start Pagination start (optional, default to 0)
      * @param  int|null $limit Items shown per page (optional)
+     * @param  \Pipedrive\versions\v1\Model\NumberBooleanDefault0|null $include_body Whether to include the mail message body content in the response. &#x60;0&#x60; &#x3D; Don&#39;t include, &#x60;1&#x60; &#x3D; Include. (optional)
      *
      * @throws ApiException on non-2xx response
      * @throws InvalidArgumentException|GuzzleException
      * @return \Pipedrive\versions\v1\Model\ListMailMessagesResponse
      */
-    public function getDealMailMessages($id, $start = 0, $limit = null)
+    public function getDealMailMessages($id, $start = 0, $limit = null, $include_body = null)
     {
-        list($response) = $this->getDealMailMessagesWithHttpInfo($id, $start, $limit);
+        list($response) = $this->getDealMailMessagesWithHttpInfo($id, $start, $limit, $include_body);
         return $response;
     }
 
@@ -3857,14 +3858,15 @@ class DealsApi
      * @param  int $id The ID of the deal (required)
      * @param  int|0 $start Pagination start (optional, default to 0)
      * @param  int|null $limit Items shown per page (optional)
+     * @param  \Pipedrive\versions\v1\Model\NumberBooleanDefault0|null $include_body Whether to include the mail message body content in the response. &#x60;0&#x60; &#x3D; Don&#39;t include, &#x60;1&#x60; &#x3D; Include. (optional)
      *
      * @throws ApiException on non-2xx response
      * @throws InvalidArgumentException|GuzzleException
      * @return array<mixed> of \Pipedrive\versions\v1\Model\ListMailMessagesResponse, HTTP status code, HTTP response headers (array of strings)
      */
-    public function getDealMailMessagesWithHttpInfo($id, $start = 0, $limit = null)
+    public function getDealMailMessagesWithHttpInfo($id, $start = 0, $limit = null, $include_body = null)
     {
-        $request = $this->getDealMailMessagesRequest($id, $start, $limit);
+        $request = $this->getDealMailMessagesRequest($id, $start, $limit, $include_body);
 
         try {
             $options = $this->createHttpClientOption();
@@ -3873,7 +3875,7 @@ class DealsApi
             } catch (RequestException $e) {
                 if ($e->getCode() === 401 && $this->config->isRefreshPossible()) {
                     $this->config->refreshToken();
-                    $request = $this->getDealMailMessagesRequest($id, $start, $limit);
+                    $request = $this->getDealMailMessagesRequest($id, $start, $limit, $include_body);
                     $response = $this->client->send($request, $options);
                 } else {
                     throw new ApiException(
@@ -3960,13 +3962,14 @@ class DealsApi
      * @param  int $id The ID of the deal (required)
      * @param  int|0 $start Pagination start (optional, default to 0)
      * @param  int|null $limit Items shown per page (optional)
+     * @param  \Pipedrive\versions\v1\Model\NumberBooleanDefault0|null $include_body Whether to include the mail message body content in the response. &#x60;0&#x60; &#x3D; Don&#39;t include, &#x60;1&#x60; &#x3D; Include. (optional)
      *
      * @throws InvalidArgumentException|OAuthProviderException
      * @return PromiseInterface
      */
-    public function getDealMailMessagesAsync($id, $start = 0, $limit = null): PromiseInterface
+    public function getDealMailMessagesAsync($id, $start = 0, $limit = null, $include_body = null): PromiseInterface
     {
-        return $this->getDealMailMessagesAsyncWithHttpInfo($id, $start, $limit)
+        return $this->getDealMailMessagesAsyncWithHttpInfo($id, $start, $limit, $include_body)
             ->then(
                 function ($response) {
                     return $response[0];
@@ -3982,14 +3985,15 @@ class DealsApi
      * @param  int $id The ID of the deal (required)
      * @param  int|0 $start Pagination start (optional, default to 0)
      * @param  int|null $limit Items shown per page (optional)
+     * @param  \Pipedrive\versions\v1\Model\NumberBooleanDefault0|null $include_body Whether to include the mail message body content in the response. &#x60;0&#x60; &#x3D; Don&#39;t include, &#x60;1&#x60; &#x3D; Include. (optional)
      *
      * @throws InvalidArgumentException|OAuthProviderException
      * @return PromiseInterface
      */
-    public function getDealMailMessagesAsyncWithHttpInfo($id, $start = 0, $limit = null): PromiseInterface
+    public function getDealMailMessagesAsyncWithHttpInfo($id, $start = 0, $limit = null, $include_body = null): PromiseInterface
     {
         $returnType = '\Pipedrive\versions\v1\Model\ListMailMessagesResponse';
-        $request = $this->getDealMailMessagesRequest($id, $start, $limit);
+        $request = $this->getDealMailMessagesRequest($id, $start, $limit, $include_body);
 
         return $this->client
             ->sendAsync($request, $this->createHttpClientOption())
@@ -4031,11 +4035,12 @@ class DealsApi
      * @param  int $id The ID of the deal (required)
      * @param  int|0 $start Pagination start (optional, default to 0)
      * @param  int|null $limit Items shown per page (optional)
+     * @param  \Pipedrive\versions\v1\Model\NumberBooleanDefault0|null $include_body Whether to include the mail message body content in the response. &#x60;0&#x60; &#x3D; Don&#39;t include, &#x60;1&#x60; &#x3D; Include. (optional)
      *
      * @throws InvalidArgumentException|OAuthProviderException
      * @return Request
      */
-    public function getDealMailMessagesRequest($id, $start = 0, $limit = null): Request
+    public function getDealMailMessagesRequest($id, $start = 0, $limit = null, $include_body = null): Request
     {
         // verify the required parameter 'id' is set
         /* @phpstan-ignore-next-line */
@@ -4067,6 +4072,14 @@ class DealsApi
         }
         if ($limit !== null) {
             $queryParams['limit'] = $limit;
+        }
+        // query params
+        /* @phpstan-ignore-next-line */
+        if (is_array($include_body)) {
+            $include_body = ObjectSerializer::serializeCollection($include_body, '', true);
+        }
+        if ($include_body !== null) {
+            $queryParams['include_body'] = $include_body;
         }