+++ lib/versions/v1/Api/PersonsApi.php
@@ -2363,14 +2363,15 @@ class PersonsApi
      * @param  int $id The ID of the person (required)
      * @param  int|0 $start Pagination start (optional, default to 0)
      * @param  int|null $limit Items shown per page (optional)
+     * @param  \Pipedrive\versions\v1\Model\NumberBooleanDefault0|null $include_body Whether to include the mail message body content in the response. &#x60;0&#x60; &#x3D; Don&#39;t include, &#x60;1&#x60; &#x3D; Include. (optional)
      *
      * @throws ApiException on non-2xx response
      * @throws InvalidArgumentException|GuzzleException
      * @return \Pipedrive\versions\v1\Model\ListMailMessagesResponse
      */
-    public function getPersonMailMessages($id, $start = 0, $limit = null)
+    public function getPersonMailMessages($id, $start = 0, $limit = null, $include_body = null)
     {
-        list($response) = $this->getPersonMailMessagesWithHttpInfo($id, $start, $limit);
+        list($response) = $this->getPersonMailMessagesWithHttpInfo($id, $start, $limit, $include_body);
         return $response;
     }
 
@@ -2382,14 +2383,15 @@ class PersonsApi
      * @param  int $id The ID of the person (required)
      * @param  int|0 $start Pagination start (optional, default to 0)
      * @param  int|null $limit Items shown per page (optional)
+     * @param  \Pipedrive\versions\v1\Model\NumberBooleanDefault0|null $include_body Whether to include the mail message body content in the response. &#x60;0&#x60; &#x3D; Don&#39;t include, &#x60;1&#x60; &#x3D; Include. (optional)
      *
      * @throws ApiException on non-2xx response
      * @throws InvalidArgumentException|GuzzleException
      * @return array<mixed> of \Pipedrive\versions\v1\Model\ListMailMessagesResponse, HTTP status code, HTTP response headers (array of strings)
      */
-    public function getPersonMailMessagesWithHttpInfo($id, $start = 0, $limit = null)
+    public function getPersonMailMessagesWithHttpInfo($id, $start = 0, $limit = null, $include_body = null)
     {
-        $request = $this->getPersonMailMessagesRequest($id, $start, $limit);
+        $request = $this->getPersonMailMessagesRequest($id, $start, $limit, $include_body);
 
         try {
             $options = $this->createHttpClientOption();
@@ -2398,7 +2400,7 @@ class PersonsApi
             } catch (RequestException $e) {
                 if ($e->getCode() === 401 && $this->config->isRefreshPossible()) {
                     $this->config->refreshToken();
-                    $request = $this->getPersonMailMessagesRequest($id, $start, $limit);
+                    $request = $this->getPersonMailMessagesRequest($id, $start, $limit, $include_body);
                     $response = $this->client->send($request, $options);
                 } else {
                     throw new ApiException(
@@ -2485,13 +2487,14 @@ class PersonsApi
      * @param  int $id The ID of the person (required)
      * @param  int|0 $start Pagination start (optional, default to 0)
      * @param  int|null $limit Items shown per page (optional)
+     * @param  \Pipedrive\versions\v1\Model\NumberBooleanDefault0|null $include_body Whether to include the mail message body content in the response. &#x60;0&#x60; &#x3D; Don&#39;t include, &#x60;1&#x60; &#x3D; Include. (optional)
      *
      * @throws InvalidArgumentException|OAuthProviderException
      * @return PromiseInterface
      */
-    public function getPersonMailMessagesAsync($id, $start = 0, $limit = null): PromiseInterface
+    public function getPersonMailMessagesAsync($id, $start = 0, $limit = null, $include_body = null): PromiseInterface
     {
-        return $this->getPersonMailMessagesAsyncWithHttpInfo($id, $start, $limit)
+        return $this->getPersonMailMessagesAsyncWithHttpInfo($id, $start, $limit, $include_body)
             ->then(
                 function ($response) {
                     return $response[0];
@@ -2507,14 +2510,15 @@ class PersonsApi
      * @param  int $id The ID of the person (required)
      * @param  int|0 $start Pagination start (optional, default to 0)
      * @param  int|null $limit Items shown per page (optional)
+     * @param  \Pipedrive\versions\v1\Model\NumberBooleanDefault0|null $include_body Whether to include the mail message body content in the response. &#x60;0&#x60; &#x3D; Don&#39;t include, &#x60;1&#x60; &#x3D; Include. (optional)
      *
      * @throws InvalidArgumentException|OAuthProviderException
      * @return PromiseInterface
      */
-    public function getPersonMailMessagesAsyncWithHttpInfo($id, $start = 0, $limit = null): PromiseInterface
+    public function getPersonMailMessagesAsyncWithHttpInfo($id, $start = 0, $limit = null, $include_body = null): PromiseInterface
     {
         $returnType = '\Pipedrive\versions\v1\Model\ListMailMessagesResponse';
-        $request = $this->getPersonMailMessagesRequest($id, $start, $limit);
+        $request = $this->getPersonMailMessagesRequest($id, $start, $limit, $include_body);
 
         return $this->client
             ->sendAsync($request, $this->createHttpClientOption())
@@ -2556,11 +2560,12 @@ class PersonsApi
      * @param  int $id The ID of the person (required)
      * @param  int|0 $start Pagination start (optional, default to 0)
      * @param  int|null $limit Items shown per page (optional)
+     * @param  \Pipedrive\versions\v1\Model\NumberBooleanDefault0|null $include_body Whether to include the mail message body content in the response. &#x60;0&#x60; &#x3D; Don&#39;t include, &#x60;1&#x60; &#x3D; Include. (optional)
      *
      * @throws InvalidArgumentException|OAuthProviderException
      * @return Request
      */
-    public function getPersonMailMessagesRequest($id, $start = 0, $limit = null): Request
+    public function getPersonMailMessagesRequest($id, $start = 0, $limit = null, $include_body = null): Request
     {
         // verify the required parameter 'id' is set
         /* @phpstan-ignore-next-line */
@@ -2592,6 +2597,14 @@ class PersonsApi
         }
         if ($limit !== null) {
             $queryParams['limit'] = $limit;
+        }
+        // query params
+        /* @phpstan-ignore-next-line */
+        if (is_array($include_body)) {
+            $include_body = ObjectSerializer::serializeCollection($include_body, '', true);
+        }
+        if ($include_body !== null) {
+            $queryParams['include_body'] = $include_body;
         }